backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
pgvector==0.2.4
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
google-api-python-client==2.116.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
openai==1.10.0
tiktoken==0.5.2
langchain==0.1.4
langchain-openai==0.0.5
langchain-community==0.0.16
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytest==8.0.0
httpx==0.26.0
